Added typescript and server url env
