+++ .env.local.example
@@ -1,2 +1,3 @@
 MONGODB_DB=
 MONGODB_URI=
+NEXT_PUBLIC_SERVER_URI=
