--- pages/_app.js
@@ -1,14 +1,0 @@
-import store from "../app/store";
-import {Provider} from "react-redux";
-
-
-function MyApp({Component, pageProps}) {
-    return (
-        <Provider store={store}>
-            <Component {...pageProps} />}
-        </Provider>
-    )
-}
-
-
-export default MyApp;
