+++ pages/_app.tsx
@@ -1,0 +1,12 @@
+import store from "../app/store";
+import { Provider } from "react-redux";
+
+function MyApp({ Component, pageProps }) {
+  return (
+    <Provider store={store}>
+      <Component {...pageProps} />
+    </Provider>
+  );
+}
+
+export default MyApp;
